moved layout folder to inside pages
