+++ src/pages/layout/Header/HeaderStyles.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/variables.scss";
+@import "../../../styles/variables";
 
 .header {
   display: flex;
