+++ src/pages/layout/MainLayout.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 type MainLayoutProps = {
   children: React.ReactNode;
 };
